+++ lib/ai_client.rb
@@ -29,6 +29,7 @@ require_relative 'ai_client/middleware'
 require_relative 'ai_client/open_router_extensions'
 require_relative 'ai_client/llm' # SMELL: must come after the open router stuff
 require_relative 'ai_client/tool'
+require_relative 'ai_client/function'
 
 # Create a generic client instance using only model name
 #   client = AiClient.new('gpt-3.5-turbo')