+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## Released
 
+### [0.3.0] - 2024-10-13
+- Breaking Change
+- Added new class AiClient::Function to encapsulate the callback functions used as tools in chats.
+- Updated the examples/tools.rb file to use the new function class
+
 ### [0.2.5] - 2024-10-11
-- Added examples/tool.rb to demonstrate use of function callbacks to provide information to the LLM when it needs it.
+- Added examples/tools.rb to demonstrate use of function callbacks to provide information to the LLM when it needs it.
 
 ### [0.2.4] - 2024-10-10
 - constrained gem omniai-openai to version 1.8.3+ for access to open_router.ai